phase-II/backend/requirements.txt
# FastAPI and ASGI server
fastapi>=0.109.0
uvicorn[standard]>=0.27.0

# Database
sqlmodel>=0.0.14
asyncpg>=0.29.0
greenlet>=3.0.0

# Authentication
pyjwt>=2.8.0

# Configuration
pydantic-settings>=2.1.0
python-dotenv>=1.0.0

# Database migrations
alembic>=1.13.0

# Password hashing (for Better Auth compatibility)
bcrypt>=4.1.0

# CORS
python-multipart>=0.0.6
